+++ computer-vision/draw_detections.py
@@ -1,5 +1,5 @@
-from PIL import Image, ImageDraw, ImageFont
 import os
+import cv2
 
 def save_image(image, filename, ext):
     n = 0
@@ -9,36 +9,21 @@ def save_image(image, filename, ext):
         n += 1
     image.save("output/" + filename + "_" + str(n) + ext)
 
-def draw_objects(image_path, objects, label=False, show=False):
-    image = Image.open(image_path)
-    draw = ImageDraw.Draw(image)
-    font = ImageFont.truetype(font="assets/Hack-Bold.ttf", size=15)
+# using basic cv2 to draw lol. Can use pillow later tho
+def draw_objects(image_path, objects):
+    img = cv2.imread(image_path)
+    w = img.shape[1]
+    h = img.shape[0]
+
     for object in objects:
-        draw.rectangle((
-            object.rectangle.x,
-            object.rectangle.y,
-            object.rectangle.x + object.rectangle.w,
-            object.rectangle.y + object.rectangle.h
-        ), fill=None, outline="red", width=2)
-        if label: draw.text((object.rectangle.x + 1, object.rectangle.y + 1), object.object_property, font=font, fill="red")
-    if show: image.show()
+        start = (int(object.bounding_box.left*w), int(object.bounding_box.top*h))
+        end = (int(object.bounding_box.left*w + object.bounding_box.width*w), int(object.bounding_box.top*h + object.bounding_box.height*h))
+        color = (0, 0, 255)
+        thickness = 2
+        img = cv2.rectangle(img, start, end, color, thickness)
+    cv2.imshow(image_path, img)
+    cv2.waitKey(0) 
+    cv2.destroyAllWindows()
 
-    filename, ext = os.path.splitext(os.path.basename(image_path))
-    save_image(image, filename, ext)
-
-def draw_faces(image_path, faces, show=False):
-    image = Image.open(image_path)
-    draw = ImageDraw.Draw(image)
-    font = ImageFont.truetype(font="assets/Hack-Bold.ttf", size=15)
-    for face in faces:
-        draw.rectangle((
-            face.face_rectangle.left,
-            face.face_rectangle.top,
-            face.face_rectangle.left + face.face_rectangle.width,
-            face.face_rectangle.top + face.face_rectangle.height
-        ), fill=None, outline="red", width=2)
-        # draw.text((face.face_rectangle.left + 1, face.face_rectangle.top + 1), "face", font=font, fill="red")
-    if show: image.show()
-
-    filename, ext = os.path.splitext(os.path.basename(image_path))
-    save_image(image, filename, ext)
+    # filename, ext = os.path.splitext(os.path.basename(image_path))
+    # save_image(image, filename, ext)
